backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pyjwt==2.8.0
bcrypt==4.1.3
pydantic==2.7.4
